Revert "Вынес set prioritizedTasks и всю логику работы с ним в отдельный класс PriorityTaskManager"
This reverts commit 065205cb436ac68f8a09f4656564d1e32ff5446c.
