--- src/service/interfaces/PriorityManager.java
@@ -1,16 +1,0 @@
-package service.interfaces;
-
-import model.Task;
-
-import java.util.List;
-
-public interface PriorityManager {
-
-    boolean addTask(Task taskToAdd);
-
-    boolean removeTask(Task taskToRemove);
-
-    boolean containsTask(Task taskToCheck);
-
-    List<Task> getPrioritizedTasks();
-}
